Minor changes to the model/entity files
This commit makes minor changes to the model/entity files to allow failing tests pass.
These changes are described below.
    - changes data type of quantityReplenished in the StockReplenishment and StockReplenishmentEntity classes from int to long
    - removes unused imports in StockLevelEntity class. Also adds the Setter lombok annotation

+++ src/main/java/co/siegerand/stocklevelservice/model/StockReplenishment.java
@@ -13,7 +13,7 @@ import co.siegerand.stocklevelservice.persistence.entity.StockReplenishmentEntit
 public class StockReplenishment {
 
     private final int bookId;
-    private final int quantityReplenished;
+    private final long quantityReplenished;
     private final ZonedDateTime timestamp;
 
     public StockReplenishment(StockReplenishmentEntity entity) {
